Cat vs Dog Classification project/requirements.txt
tensorflow>=2.15.0
streamlit
pillow
numpy
matplotlib